% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_function.R
\name{likert_labelled}
\alias{likert_labelled}
\title{Produces likert plots with labels}
\usage{
likert_labelled(ds, mid_level, ranked_levels, group,
  label_minimum = 0.05, group_arrange = "positive")
}
\arguments{
\item{ds}{a datafarme with groups as rows and levels as columns}

\item{ranked_levels}{an ordered factor of all levels}

\item{group}{a factor or character vector of all groups (y-axis of likert plot)}

\item{label_minimum}{the proportion for which likert plot will be labelled}

\item{group_arrange}{a character of positive (default) or negative}

\item{mid_ldevel}{the neutral or middle category}
}
\value{
None
}
\description{
The one main function for the \code{\link{precisionlikert}} package.
}
\examples{
data("country_outlook")
ranked_levels <- factor(x = c('Much.worse', 'Somewhat.worse', 'Almost.the.same', 'Somewhat.better', 'Much.better'), levels = c('Much.worse', 'Somewhat.worse', 'Almost.the.same', 'Somewhat.better', 'Much.better'))
mid_level <- 'Almost.the.same'
group <- 'Country'
likert_labeled(ds, mid_level, ranked_levels, group)
}
